useCart update
