+++ src/lib/cart.ts
@@ -8,3 +8,10 @@ export const transformCart = (cart: ICart): ICartBase => ({
   totalPrice: cart.totalPrice,
   status: cart.status,
 });
+
+export const truncateDescription = (description: string, maxLength: number) => {
+  if (description.length > maxLength) {
+    return description.substring(0, maxLength) + "...";
+  }
+  return description;
+};
